requirements.txt
PyYAML==6.0.2
python-hcl2==7.3.1